query hummock stats * need to query for upstream tables being scanned instead

+++ src/frontend/src/expr/table_function.rs
@@ -589,7 +589,7 @@ impl TableFunction {
             args: vec![],
             return_type: DataType::Struct(StructType::new(vec![
                 ("job_id".to_owned(), DataType::Int32),
-                ("row_count".to_owned(), DataType::Int64),
+                ("row_count".to_owned(), DataType::Decimal),
             ])),
             function_type: TableFunctionType::InternalBackfillProgress,
             user_defined: None,
